ref: create standard location for groovy scripts in package
